assignments8

+++ assignments/src/literals/ByteLiteral.java
@@ -5,9 +5,9 @@ class ByteLiteral{
 	byte sample2 = (byte)0B10000000;
 	byte sample3 = (byte)0B10000011;
 	byte sample4 = 0B01111111; //implicit
-	System.out.println("must be -1: " + (int)sample);
-	System.out.println("must be -128: " + (int)sample2);
-	System.out.println("must be -125: " + (int)sample3);
-	System.out.println("must be 127: " + (int)sample4);	
+	System.out.println("must be -1: " + sample);
+	System.out.println("must be -128: " + sample2);
+	System.out.println("must be -125: " + sample3);
+	System.out.println("must be 127: " + sample4);	
     }
 }
